# 03-todo-api/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateTaskRequest:
    properties:
      description:
        type: string
      title:
        maxLength: 200
        type: string
    required:
    - title
    type: object
  models.Task:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - title
    type: object
  models.TasksResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      total:
        type: integer
    type: object
  models.UpdateTaskRequest:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        maxLength: 200
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API для управления задачами (To-Do List)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Todo List API
  version: "1.0"
paths:
  /tasks:
    get:
      consumes:
      - application/json
      description: Возвращает список задач с поддержкой пагинации, фильтрации и поиска
      parameters:
      - description: Лимит (по умолчанию 10)
        in: query
        name: limit
        type: integer
      - description: Смещение (по умолчанию 0)
        in: query
        name: offset
        type: integer
      - description: Фильтр по статусу выполнения
        in: query
        name: completed
        type: boolean
      - description: Поиск по заголовку и описанию
        in: query
        name: search
        type: string
      - description: Поле для сортировки (created_at, completed)
        in: query
        name: sort_by
        type: string
      - description: Порядок сортировки (asc, desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TasksResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создает новую задачу с указанным заголовком и описанием
      parameters:
      - description: Данные для создания задачи
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новую задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет задачу по её ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить задачу
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Возвращает задачу по её уникальному идентификатору
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить задачу по ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Обновляет данные задачи по её ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить задачу
      tags:
      - tasks
  /tasks/{id}/complete:
    patch:
      consumes:
      - application/json
      description: Устанавливает статус выполнения задачи в true
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отметить задачу как выполненную
      tags:
      - tasks
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"